+++ cx-portal/src/components/pages/Admin/components/RegistrationRequests/index.tsx
@@ -29,7 +29,7 @@ import CompanyDetailOverlay from './CompanyDetailOverlay'
 import ConfirmationOverlay from './ConfirmationOverlay/ConfirmationOverlay'
 import {
   useApproveRequestMutation,
-  useDeclineRequestMutation,
+  useDeclineChecklistMutation,
   useFetchCompanySearchQuery,
   useFetchDocumentByIdMutation,
   useUpdateBPNMutation,
@@ -56,7 +56,7 @@ export default function RegistrationRequests() {
   const [actionType, setActionType] = useState<string>('approve')
 
   const [approveRequest] = useApproveRequestMutation()
-  const [declineRequest] = useDeclineRequestMutation()
+  const [declineRequest] = useDeclineChecklistMutation()
   const [getDocumentById] = useFetchDocumentByIdMutation()
 
   const [updateBpn] = useUpdateBPNMutation()
@@ -81,6 +81,7 @@ export default function RegistrationRequests() {
   const onTableCellClick = (params: GridCellParams) => {
     // Show overlay only when detail field clicked
     if (params.field === 'detail') {
+      setSelectedRequestId(params.row.applicationId)
       dispatch(fetchCompanyDetail(params.row.applicationId))
       setOverlayOpen(true)
     }
@@ -168,6 +169,7 @@ export default function RegistrationRequests() {
       <CompanyDetailOverlay
         {...{
           openDialog: overlayOpen,
+          selectedRequestId: selectedRequestId,
           handleOverlayClose: () => setOverlayOpen(false),
         }}
       />